Signup test implemented
